[CHANGED]: - SourceText RenderingHints to SpeedUp things a little. - Disabled ffmpeg switch for outputs under Ubuntu (Avconv is Ubuntu 14.04 default - No ffmpeg provided anymore by default) - AudioSource gstreamer commands ported to 1.x version. - SpeakerOut gstreamer command ported to 1.x version. - Text Stream-Panel now integrates QRCode functions. - SourceMovie GStreamer command ported to 1.x to be compatible with Ubutnu 14.04 h264 playback. (No more gstreamer0.10-ffmpeg package in the default repos) - Stream-Desktop now match actual system screen size when added. (Thanks Patrick Balleux for screen.java) - Ported rtmp rtsp and http URL Stream-Panel to Gstreamer 1.x (Thanks RJ)
[ADDED]:
- AudioSource Stream-Panel (Combine Mic + Monitor + AudioSources all in one) (Thanks Patrick Balleux for PaCTL.java and Joshua Lee for the request)

[REMOVED]:
- Microphone and SoundMonitor Stream-Panels
- QRCode Stream-Panel

[FIXED]:
- SourceChannel Apply Effect/Transition
- /trunk/src/com/jhlabs/math/Noise.java from Netbeans Refactoring
- Fixed Output recording to mp4 files in Ubuntu 14.04 ( -strict experimental)
- "gstreamer1.0-tools" added to control debian dependeces.
- rewrited ffmpeg/avconv BroadcastHQ/Red5HQ commands (didn't show video, only audio ... Red5 still not tested !!!)

[TODO]:
- "WaterFx" effect to be add in save/load operations ...
- rtmp, rtsp and http commands in Gstreamer 1.x format with effectv.
- "Playlist Editor for WS" (Thanks Apollo FM)
- Have to fix webcam count if writing errors occurs ...
- BackEnds AutoSwitch based on distro's default backends availability.
- Technically, ProcessRenderer should be modified to launch each existing
process and then Capturer should read from the InputStream of the process
instead of reading from a local TCPIP port. (Patrick Balleux)

+++ trunk/src/com/jhlabs/math/Noise.java
@@ -90,7 +90,7 @@ public class Noise implements Function1D, Function2D, Function3D {
             t = x + N;
             bx0 = ((int)t) & BM;
             bx1 = (bx0+1) & BM;
-            rx0 = t - t;
+            rx0 = t - (int)t;
             rx1 = rx0 - 1.0f;
             
             sx = sCurve(rx0);
@@ -119,13 +119,13 @@ public class Noise implements Function1D, Function2D, Function3D {
         t = x + N;
         bx0 = ((int)t) & BM;
         bx1 = (bx0+1) & BM;
-        rx0 = t - t;
+	rx0 = t - (int)t;
         rx1 = rx0 - 1.0f;
         
         t = y + N;
         by0 = ((int)t) & BM;
         by1 = (by0+1) & BM;
-        ry0 = t - t;
+	ry0 = t - (int)t;
         ry1 = ry0 - 1.0f;
         
         i = p[bx0];
@@ -171,19 +171,19 @@ public class Noise implements Function1D, Function2D, Function3D {
         t = x + N;
         bx0 = ((int)t) & BM;
         bx1 = (bx0+1) & BM;
-        rx0 = t - t;
+	rx0 = t - (int)t;
         rx1 = rx0 - 1.0f;
         
         t = y + N;
         by0 = ((int)t) & BM;
         by1 = (by0+1) & BM;
-        ry0 = t - t;
+	ry0 = t - (int)t;
         ry1 = ry0 - 1.0f;
         
         t = z + N;
         bz0 = ((int)t) & BM;
         bz1 = (bz0+1) & BM;
-        rz0 = t - t;
+	rz0 = t - (int)t;
         rz1 = rz0 - 1.0f;
         
         i = p[bx0];
